Initial complete state handling for rollout mode
- return immediately if added deployment in complete state
- move methods borrowed from controller/utils to utils
- move rich info methods to deployment_debug
- DeploymentFeed.Added receives `completed` argument
